+++ Source/FileSystem.Tests/FileSystem.Tests.csproj
@@ -3,16 +3,15 @@
 	<PropertyGroup>
 		<TargetFramework>net6.0</TargetFramework>
 		<Nullable>enable</Nullable>
-
 		<IsPackable>false</IsPackable>
 	</PropertyGroup>
 
 	<ItemGroup>
 		<PackageReference Include="FluentAssertions" Version="6.4.0" />
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.11.0" />
+		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.0.0" />
 		<PackageReference Include="Moq" Version="4.16.1" />
 		<PackageReference Include="morelinq" Version="3.3.2" />
-		<PackageReference Include="System.IO.Abstractions.TestingHelpers" Version="16.1.4" />
+		<PackageReference Include="System.IO.Abstractions.TestingHelpers" Version="16.1.7" />
 		<PackageReference Include="System.Reactive" Version="5.0.0" />
 		<PackageReference Include="xunit" Version="2.4.1" />
 		<PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
