+++ Source/Network/Network.csproj
@@ -8,7 +8,7 @@
 
 	<ItemGroup>
 		<PackageReference Include="Microsoft.Extensions.Http" Version="6.0.0" />
-		<PackageReference Include="System.IO.Abstractions" Version="13.2.47" />
+		<PackageReference Include="System.IO.Abstractions" Version="16.1.7" />
 	</ItemGroup>
 
 	<ItemGroup>
