+++ Source/Reflection/Reflection.csproj
@@ -7,7 +7,7 @@
 	<Import Project="..\..\Build\Common.props" />
 
 	<ItemGroup>
-		<PackageReference Include="CSharpFunctionalExtensions" Version="2.24.3" />
+		<PackageReference Include="CSharpFunctionalExtensions" Version="2.27.2" />
 	</ItemGroup>
 
 </Project>
