+++ Source/XPath.Tests/XPath.Tests.csproj
@@ -8,7 +8,7 @@
 	<Import Project="..\..\Build\Common.props" />
 
 	<ItemGroup>
-		<PackageReference Include="FluentAssertions" Version="6.2.0" />
+		<PackageReference Include="FluentAssertions" Version="6.4.0" />
 		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.0.0" />
 		<PackageReference Include="xunit" Version="2.4.1" />
 		<PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
